+++ lib/prf.dart
@@ -69,10 +69,10 @@ export 'prf_types/legacy/prf_string.dart';
 export 'prf_types/legacy/prf_string_list.dart';
 export 'prf_types/legacy/prf_theme_mode.dart';
 export 'prf_types/prf.dart';
-export 'prfy_types/prf_enum.dart';
-export 'prfy_types/prf_json.dart';
-export 'prfy_types/prfy.dart';
-export 'prfy_types/prfy_enum.dart';
-export 'prfy_types/prfy_json.dart';
+export 'prf_types/legacy/prf_enum.dart';
+export 'prf_types/legacy/prf_json.dart';
+export 'prf_types/legacy/prfy.dart';
+export 'prf_types/legacy/prfy_enum.dart';
+export 'prf_types/legacy/prfy_json.dart';
 export 'services/prf_cooldown.dart';
 export 'services/prf_rate_limiter.dart';
